feat: oracle dashboard card aligns
